Rename `csearch.rs` -> `cstore_impl.rs`.
