+++ src/librustc_metadata/cstore_impl.rs
